+++ vello_tests/snapshots/smoke/gradient_color_alpha_premultiplied.png
