+++ vello_tests/tests/regression.rs
@@ -8,7 +8,9 @@ use scenes::SimpleText;
 use vello::{
     AaConfig, Scene,
     kurbo::{Affine, Rect, RoundedRect, Stroke},
-    peniko::{Extend, ImageQuality, color::palette},
+    peniko::{
+        Color, ColorStop, Extend, Gradient, ImageQuality, InterpolationAlphaSpace, color::palette,
+    },
 };
 use vello_tests::{TestParams, smoke_snapshot_test_sync, snapshot_test_sync};
 
@@ -138,6 +140,70 @@ fn text_stroke_width_zero() {
         (font_size * 1.25).ceil() as _,
     );
     snapshot_test_sync(scene, &params)
+        .unwrap()
+        .assert_mean_less_than(0.001);
+}
+
+/// <https://github.com/web-platform-tests/wpt/blob/18c64a74b1/html/canvas/element/fill-and-stroke-styles/2d.gradient.interpolate.coloralpha.html>
+/// See <https://github.com/linebender/vello/issues/1056>.
+#[test]
+#[cfg_attr(skip_gpu_tests, ignore)]
+fn test_gradient_color_alpha_premultiplied() {
+    let mut scene = Scene::new();
+    let viewport = Rect::new(0., 0., 100., 50.);
+    scene.fill(
+        vello::peniko::Fill::NonZero,
+        Affine::IDENTITY,
+        &Gradient::new_linear((0., 0.), (100., 0.))
+            .with_stops([
+                ColorStop {
+                    offset: 0.,
+                    color: Color::from_rgba8(255, 255, 0, 0).into(),
+                },
+                ColorStop {
+                    offset: 1.,
+                    color: Color::from_rgba8(0, 0, 255, 255).into(),
+                },
+            ])
+            .with_interpolation_alpha_space(InterpolationAlphaSpace::Premultiplied),
+        None,
+        &viewport,
+    );
+    let mut params = TestParams::new("gradient_color_alpha_premultiplied", 100, 50);
+    params.base_color = Some(palette::css::WHITE);
+    smoke_snapshot_test_sync(scene, &params)
+        .unwrap()
+        .assert_mean_less_than(0.001);
+}
+
+/// <https://github.com/web-platform-tests/wpt/blob/18c64a74b1/html/canvas/element/fill-and-stroke-styles/2d.gradient.interpolate.coloralpha.html>
+/// See <https://github.com/linebender/vello/issues/1056>.
+#[test]
+#[cfg_attr(skip_gpu_tests, ignore)]
+fn test_gradient_color_alpha_unpremultiplied() {
+    let mut scene = Scene::new();
+    let viewport = Rect::new(0., 0., 100., 50.);
+    scene.fill(
+        vello::peniko::Fill::NonZero,
+        Affine::IDENTITY,
+        &Gradient::new_linear((0., 0.), (100., 0.))
+            .with_stops([
+                ColorStop {
+                    offset: 0.,
+                    color: Color::from_rgba8(255, 255, 0, 0).into(),
+                },
+                ColorStop {
+                    offset: 1.,
+                    color: Color::from_rgba8(0, 0, 255, 255).into(),
+                },
+            ])
+            .with_interpolation_alpha_space(InterpolationAlphaSpace::Unpremultiplied),
+        None,
+        &viewport,
+    );
+    let mut params = TestParams::new("gradient_color_alpha_unpremultiplied", 100, 50);
+    params.base_color = Some(palette::css::WHITE);
+    smoke_snapshot_test_sync(scene, &params)
         .unwrap()
         .assert_mean_less_than(0.001);
 }
